Fix inconsistency of the modification time of auto-saved file.
`do-auto-save' does not update the record of the last modification
time of the file being visited by the current buffer but it should
do update since we know that the auto-saved file is the file being
visited by the buffer.  Use `clear-visited-file-modtime' instead of
updating the record of the modification time since there seems to
be no interface to change the record.